Tested the codebase with clang++ -Wall and gcc -Wall. Corrected numerous issues:
- unused variables
- variables that are used before initialized.
- reordering of initializer lists
- symbols declared as class in one place and (forward) declared as struct somewhere else.
- missing virtual destructors.
- comparison between signed and unsigned integer expressions

Option -Wall is now default compiler flag.

+++ src/precice/impl/SolverInterfaceImpl.hpp
@@ -286,7 +286,7 @@ public:
    */
   void getMeshVertices (
     int     meshID,
-    int     size,
+    size_t  size,
     int*    ids,
     double* positions );
 
@@ -299,7 +299,7 @@ public:
    */
   void getMeshVertexIDsFromPositions (
     int     meshID,
-    int     size,
+    size_t  size,
     double* positions,
     int*    ids );
